refactor: move shard splitter into controller
